+++ packages/orchestration/src/examples/swapExample.contract.js
@@ -1,17 +1,11 @@
-import { prepareAsyncFlowTools } from '@agoric/async-flow';
 import { StorageNodeShape } from '@agoric/internal';
 import { TimerServiceShape } from '@agoric/time';
-import { prepareVowTools } from '@agoric/vow';
-import { prepareRecorderKitMakers } from '@agoric/zoe/src/contractSupport/recorder.js';
 import { withdrawFromSeat } from '@agoric/zoe/src/contractSupport/zoeHelpers.js';
-import { makeDurableZone } from '@agoric/zone/durable.js';
 import { Far } from '@endo/far';
 import { deeplyFulfilled } from '@endo/marshal';
 import { M, objectMap } from '@endo/patterns';
-import { prepareLocalChainAccountKit } from '../exos/local-chain-account-kit.js';
-import { makeOrchestrationFacade } from '../facade.js';
-import { makeChainHub } from '../utils/chainHub.js';
 import { orcUtils } from '../utils/orc.js';
+import { provideOrchestration } from '../utils/start-helper.js';
 
 /**
  * @import {Orchestrator, IcaAccount, CosmosValidatorAddress} from '../types.js'
@@ -59,10 +53,6 @@ export const makeNatAmountShape = (brand, min) =>
  * @param {Baggage} baggage
  */
 export const start = async (zcf, privateArgs, baggage) => {
-  const { brands } = zcf.getTerms();
-
-  const zone = makeDurableZone(baggage);
-
   const {
     agoricNames,
     localchain,
@@ -72,31 +62,20 @@ export const start = async (zcf, privateArgs, baggage) => {
     marshaller,
   } = privateArgs;
 
-  const chainHub = makeChainHub(agoricNames);
-  const { makeRecorderKit } = prepareRecorderKitMakers(baggage, marshaller);
-  const makeLocalChainAccountKit = prepareLocalChainAccountKit(
-    zone,
-    makeRecorderKit,
+  const { orchestrate } = provideOrchestration(
     zcf,
-    timerService,
-    chainHub,
+    baggage,
+    {
+      agoricNames,
+      localchain,
+      orchestrationService,
+      storageNode,
+      timerService,
+    },
+    marshaller,
   );
-  const vowTools = prepareVowTools(zone.subZone('vows'));
-  const asyncFlowTools = prepareAsyncFlowTools(zone.subZone('asyncFlow'), {
-    vowTools,
-  });
 
-  const { orchestrate } = makeOrchestrationFacade({
-    localchain,
-    orchestrationService,
-    storageNode,
-    timerService,
-    zcf,
-    zone,
-    chainHub,
-    makeLocalChainAccountKit,
-    asyncFlowTools,
-  });
+  const { brands } = zcf.getTerms();
 
   /** deprecated historical example */
   /**
